+++ common/src/main/java/net/spell_engine/SpellEngineMod.java
@@ -38,6 +38,7 @@ import net.spell_engine.misc.criteria.SpellCastCriteria;
 import net.spell_engine.internals.delivery.SpellStashHelper;
 import net.spell_engine.network.ServerNetwork;
 import net.spell_engine.rpg_series.RPGSeriesCore;
+import net.spell_engine.rpg_series.loot.AffiliationGroupEntry;
 import net.spell_engine.spellbinding.*;
 import net.spell_engine.spellbinding.spellchoice.SpellChoiceFeature;
 import net.spell_engine.spellbinding.spellchoice.SpellChoiceScreenHandler;
@@ -129,6 +130,7 @@ public class SpellEngineMod {
         Registry.register(Registries.BLOCK_ENTITY_TYPE, SpellBinding.ID, SpellBindingBlockEntity.ENTITY_TYPE);
         Registry.register(Registries.SCREEN_HANDLER, SpellBinding.ID, SpellBindingScreenHandler.HANDLER_TYPE);
         Registry.register(Registries.LOOT_FUNCTION_TYPE, SpellBindRandomlyLootFunction.ID, SpellBindRandomlyLootFunction.TYPE);
+        Registry.register(Registries.LOOT_POOL_ENTRY_TYPE, AffiliationGroupEntry.ID, AffiliationGroupEntry.TYPE);
         Registry.register(Registries.SCREEN_HANDLER, SpellChoiceFeature.ID, SpellChoiceScreenHandler.HANDLER_TYPE);
     }